requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
ollama==0.1.0
python-binance==1.0.21